Revert "Ajout d'un captcha (#769)"
This reverts commit e4ed50f9a355b584302ec8ed051ba72826bb78f1.
