+++ .env.app.example
@@ -41,11 +41,6 @@ DJANGO_EMAIL_USE_SSL=1
 RNB_SEND_ADDRESS=ne-pas-repondre@rnb.beta.gouv.fr
 RNB_REPLY_TO_ADDRESS=rnb@beta.gouv.fr
 
-# Captcha
-ENABLE_CAPTCHA=false
-PRIVATE_CAPTCHA_API_KEY="xxx"
-PRIVATE_CAPTCHA_SITEKEY="xxx"
-
 # WEBHOOKS
 MATTERMOST_RNB_TECH_WEBHOOK_URL=https://mattermost.example.net/hooks/xyz
 MATTERMOST_NOTIFICATIONS=0
