+++ app/app/settings.py
@@ -22,9 +22,6 @@ from app.schedule import get_celery_beat_schedule
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 WRITABLE_DATA_DIR = os.path.join(BASE_DIR, "data")
 
-PRIVATE_CAPTCHA_API_KEY = os.environ.get("PRIVATE_CAPTCHA_API_KEY")
-PRIVATE_CAPTCHA_SITEKEY = os.environ.get("PRIVATE_CAPTCHA_SITEKEY")
-ENABLE_CAPTCHA = os.environ.get("ENABLE_CAPTCHA", default="true").lower() == "true"
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/3.0/howto/deployment/checklist/
@@ -190,7 +187,6 @@ AUTH_PASSWORD_VALIDATORS = [
     },
 ]
 
-
 REST_FRAMEWORK = {
     "DEFAULT_AUTHENTICATION_CLASSES": [
         "rest_framework.authentication.TokenAuthentication"
@@ -209,9 +205,6 @@ REST_FRAMEWORK = {
     },
     "URL_FORMAT_OVERRIDE": None,
 }
-
-if ENVIRONMENT == "test":
-    REST_FRAMEWORK["DEFAULT_THROTTLE_RATES"]["create_user"] = "1000/day"  # type: ignore[index]
 
 
 SPECTACULAR_SETTINGS = {
